Post functionalities, investment activities added

+++ src/main/java/com/inbest/backend/repository/PostRepository.java
@@ -3,6 +3,6 @@ package com.inbest.backend.repository;
 import com.inbest.backend.model.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post, Integer>
+public interface PostRepository extends JpaRepository<Post, Long>
 {
 }
